backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
faiss-cpu==1.7.4
sentence-transformers==2.3.1
numpy==1.26.3
requests==2.31.0
